+++ portlets/music-brainz-portlet/docroot/WEB-INF/service/com/jkamcc/musicbrainz/service/persistence/ArtistUtil.java
@@ -133,8 +133,7 @@ public class ArtistUtil {
 	* @param id the primary key for the new artist
 	* @return the new artist
 	*/
-	public static com.jkamcc.musicbrainz.model.Artist create(
-		java.lang.Integer id) {
+	public static com.jkamcc.musicbrainz.model.Artist create(long id) {
 		return getPersistence().create(id);
 	}
 
@@ -146,8 +145,7 @@ public class ArtistUtil {
 	* @throws com.jkamcc.musicbrainz.NoSuchArtistException if a artist with the primary key could not be found
 	* @throws SystemException if a system exception occurred
 	*/
-	public static com.jkamcc.musicbrainz.model.Artist remove(
-		java.lang.Integer id)
+	public static com.jkamcc.musicbrainz.model.Artist remove(long id)
 		throws com.jkamcc.musicbrainz.NoSuchArtistException,
 			com.liferay.portal.kernel.exception.SystemException {
 		return getPersistence().remove(id);
@@ -167,8 +165,7 @@ public class ArtistUtil {
 	* @throws com.jkamcc.musicbrainz.NoSuchArtistException if a artist with the primary key could not be found
 	* @throws SystemException if a system exception occurred
 	*/
-	public static com.jkamcc.musicbrainz.model.Artist findByPrimaryKey(
-		java.lang.Integer id)
+	public static com.jkamcc.musicbrainz.model.Artist findByPrimaryKey(long id)
 		throws com.jkamcc.musicbrainz.NoSuchArtistException,
 			com.liferay.portal.kernel.exception.SystemException {
 		return getPersistence().findByPrimaryKey(id);
@@ -181,8 +178,7 @@ public class ArtistUtil {
 	* @return the artist, or <code>null</code> if a artist with the primary key could not be found
 	* @throws SystemException if a system exception occurred
 	*/
-	public static com.jkamcc.musicbrainz.model.Artist fetchByPrimaryKey(
-		java.lang.Integer id)
+	public static com.jkamcc.musicbrainz.model.Artist fetchByPrimaryKey(long id)
 		throws com.liferay.portal.kernel.exception.SystemException {
 		return getPersistence().fetchByPrimaryKey(id);
 	}
